i64

+++ fastestserver/src/db.rs
@@ -3,7 +3,7 @@ use serde;
 #[derive(Debug, Clone)]
 pub struct LevelUpdate {
     pub new_level: i32,
-    pub time_at: i32,
+    pub time_at: i64,
     pub username: String,
 }
 
@@ -14,7 +14,7 @@ pub struct RankUpdate {
     pub new_rank: i32,
     pub old_rank_name: Option<String>,
     pub new_rank_name: String,
-    pub time_at: i32,
+    pub time_at: i64,
     pub rank_type: String,
 }
 
@@ -25,7 +25,7 @@ pub struct PartialRankUpdate {
     pub new_rank: i32,
     // pub old_rank_name: Option<String>,
     pub new_rank_name: String,
-    pub time_at: i32,
+    pub time_at: i64,
     // pub rank_type: String,
 }
 
@@ -33,15 +33,15 @@ pub struct PartialRankUpdate {
 pub struct PlayingTime {
     pub username: String,
     pub gamename: String,
-    pub started_at: i32,
-    pub ended_at: i32,
+    pub started_at: i64,
+    pub ended_at: i64,
 }
 
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct PlayingNow {
     pub username: String,
     pub gamename: String,
-    pub started_at: i32,
+    pub started_at: i64,
 }
 
 #[derive(Debug, Clone, serde::Serialize)]
@@ -50,5 +50,5 @@ pub struct MonthlyCheck {
     pub gamename: String,
     pub month: i32,
     pub year: i32,
-    pub playtime: i32, // in seconds
+    pub playtime: i64, // in seconds
 }
